Update Question Answers Entities and DTOs to support returning multiple answers per question.
OASys supports check box lists which have multiple questions for a single question. The existing implementation only allowed a single answer per question.

Where specific answers are hardcoded into the API (e.g. for Needs and Safeguarding risk flags) it is known that there is a maximum of 1 answer for those questions which is reflected in the code using getOrNull(0)